Simplify HTTP/1.1 state machine
Buffer management should be a responsibility of the caller, whereas
the state machine simply tells how much it processed. If it couldn't
advance (i.e. if there was an error or if there simply wasn't enough
data to do anything) the caller receives an error. The caller is
expected to provide enough data in a contiguous buffer.

In future, this type should be parametrised on functions that handle
a complete Request and the request body data.