request: Use offsets for parsed strings
This reduces the size of the Request, and in future, will make it
easier to copy the Request (simply dupe the data buffer).

This commit also simplifies the header field map (partly to make it
easier to for implementing offset pointers) to just use a naive
linear lookup on a slice. This may even be faster than any fancier hash
map, but still needs more in-depth benchmarking.